# evmrs: skip step check in tail calls