--- a/rust/fuzz/fuzz_targets/evmc_execute.rs
+++ b/rust/fuzz/fuzz_targets/evmc_execute.rs
@@ -41,16 +41,14 @@ impl<'a> Arbitrary<'a> for InterpreterArgs<'a> {
         let input = <&[u8]>::arbitrary(u)?;
         let code = <&[u8]>::arbitrary(u)?;
         let message = evmc_message {
-            kind: u
-                .choose(&[
-                    MessageKind::EVMC_CALL,
-                    MessageKind::EVMC_CALLCODE,
-                    MessageKind::EVMC_CREATE,
-                    MessageKind::EVMC_CREATE2,
-                    MessageKind::EVMC_DELEGATECALL,
-                    MessageKind::EVMC_EOFCREATE,
-                ])?
-                .clone(),
+            kind: *u.choose(&[
+                MessageKind::EVMC_CALL,
+                MessageKind::EVMC_CALLCODE,
+                MessageKind::EVMC_CREATE,
+                MessageKind::EVMC_CREATE2,
+                MessageKind::EVMC_DELEGATECALL,
+                MessageKind::EVMC_EOFCREATE,
+            ])?,
             flags: u32::arbitrary(u)?,
             depth: i32::arbitrary(u)?,
             gas: u.int_in_range(0..=5_000_000_000)?, // see go/ct/evm_fuzz_test.go
@@ -94,20 +92,17 @@ impl<'a> Arbitrary<'a> for InterpreterArgs<'a> {
         context
             .expect_get_storage()
             .return_const(Uint256::from(u256::arbitrary(u)?));
-        context.expect_set_storage().return_const(
-            u.choose(&[
-                StorageStatus::EVMC_STORAGE_ASSIGNED,
-                StorageStatus::EVMC_STORAGE_ADDED,
-                StorageStatus::EVMC_STORAGE_DELETED,
-                StorageStatus::EVMC_STORAGE_MODIFIED,
-                StorageStatus::EVMC_STORAGE_DELETED_ADDED,
-                StorageStatus::EVMC_STORAGE_MODIFIED_DELETED,
-                StorageStatus::EVMC_STORAGE_DELETED_RESTORED,
-                StorageStatus::EVMC_STORAGE_ADDED_DELETED,
-                StorageStatus::EVMC_STORAGE_MODIFIED_RESTORED,
-            ])?
-            .clone(),
-        );
+        context.expect_set_storage().return_const(*u.choose(&[
+            StorageStatus::EVMC_STORAGE_ASSIGNED,
+            StorageStatus::EVMC_STORAGE_ADDED,
+            StorageStatus::EVMC_STORAGE_DELETED,
+            StorageStatus::EVMC_STORAGE_MODIFIED,
+            StorageStatus::EVMC_STORAGE_DELETED_ADDED,
+            StorageStatus::EVMC_STORAGE_MODIFIED_DELETED,
+            StorageStatus::EVMC_STORAGE_DELETED_RESTORED,
+            StorageStatus::EVMC_STORAGE_ADDED_DELETED,
+            StorageStatus::EVMC_STORAGE_MODIFIED_RESTORED,
+        ])?);
         context
             .expect_get_balance()
             .return_const(Uint256::from(u256::arbitrary(u)?));
@@ -124,31 +119,29 @@ impl<'a> Arbitrary<'a> for InterpreterArgs<'a> {
             .expect_selfdestruct()
             .return_const(bool::arbitrary(u)?);
         let execution_result = ExecutionResult {
-            status_code: u
-                .choose(&[
-                    StatusCode::EVMC_SUCCESS,
-                    StatusCode::EVMC_FAILURE,
-                    StatusCode::EVMC_REVERT,
-                    StatusCode::EVMC_OUT_OF_GAS,
-                    StatusCode::EVMC_INVALID_INSTRUCTION,
-                    StatusCode::EVMC_UNDEFINED_INSTRUCTION,
-                    StatusCode::EVMC_STACK_OVERFLOW,
-                    StatusCode::EVMC_STACK_UNDERFLOW,
-                    StatusCode::EVMC_BAD_JUMP_DESTINATION,
-                    StatusCode::EVMC_INVALID_MEMORY_ACCESS,
-                    StatusCode::EVMC_CALL_DEPTH_EXCEEDED,
-                    StatusCode::EVMC_STATIC_MODE_VIOLATION,
-                    StatusCode::EVMC_PRECOMPILE_FAILURE,
-                    StatusCode::EVMC_CONTRACT_VALIDATION_FAILURE,
-                    StatusCode::EVMC_ARGUMENT_OUT_OF_RANGE,
-                    StatusCode::EVMC_WASM_UNREACHABLE_INSTRUCTION,
-                    StatusCode::EVMC_WASM_TRAP,
-                    StatusCode::EVMC_INSUFFICIENT_BALANCE,
-                    StatusCode::EVMC_INTERNAL_ERROR,
-                    StatusCode::EVMC_REJECTED,
-                    StatusCode::EVMC_OUT_OF_MEMORY,
-                ])?
-                .clone(),
+            status_code: *u.choose(&[
+                StatusCode::EVMC_SUCCESS,
+                StatusCode::EVMC_FAILURE,
+                StatusCode::EVMC_REVERT,
+                StatusCode::EVMC_OUT_OF_GAS,
+                StatusCode::EVMC_INVALID_INSTRUCTION,
+                StatusCode::EVMC_UNDEFINED_INSTRUCTION,
+                StatusCode::EVMC_STACK_OVERFLOW,
+                StatusCode::EVMC_STACK_UNDERFLOW,
+                StatusCode::EVMC_BAD_JUMP_DESTINATION,
+                StatusCode::EVMC_INVALID_MEMORY_ACCESS,
+                StatusCode::EVMC_CALL_DEPTH_EXCEEDED,
+                StatusCode::EVMC_STATIC_MODE_VIOLATION,
+                StatusCode::EVMC_PRECOMPILE_FAILURE,
+                StatusCode::EVMC_CONTRACT_VALIDATION_FAILURE,
+                StatusCode::EVMC_ARGUMENT_OUT_OF_RANGE,
+                StatusCode::EVMC_WASM_UNREACHABLE_INSTRUCTION,
+                StatusCode::EVMC_WASM_TRAP,
+                StatusCode::EVMC_INSUFFICIENT_BALANCE,
+                StatusCode::EVMC_INTERNAL_ERROR,
+                StatusCode::EVMC_REJECTED,
+                StatusCode::EVMC_OUT_OF_MEMORY,
+            ])?,
             gas_left: Arbitrary::arbitrary(u)?,
             gas_refund: Arbitrary::arbitrary(u)?,
             output: Arbitrary::arbitrary(u)?,
@@ -166,44 +159,36 @@ impl<'a> Arbitrary<'a> for InterpreterArgs<'a> {
             .expect_get_block_hash()
             .return_const(Uint256::from(u256::arbitrary(u)?));
         context.expect_emit_log().return_const(());
-        context.expect_access_account().return_const(
-            u.choose(&[
-                AccessStatus::EVMC_ACCESS_COLD,
-                AccessStatus::EVMC_ACCESS_WARM,
-            ])?
-            .clone(),
-        );
-        context.expect_access_storage().return_const(
-            u.choose(&[
-                AccessStatus::EVMC_ACCESS_COLD,
-                AccessStatus::EVMC_ACCESS_WARM,
-            ])?
-            .clone(),
-        );
+        context.expect_access_account().return_const(*u.choose(&[
+            AccessStatus::EVMC_ACCESS_COLD,
+            AccessStatus::EVMC_ACCESS_WARM,
+        ])?);
+        context.expect_access_storage().return_const(*u.choose(&[
+            AccessStatus::EVMC_ACCESS_COLD,
+            AccessStatus::EVMC_ACCESS_WARM,
+        ])?);
         context
             .expect_get_transient_storage()
             .return_const(Uint256::from(u256::arbitrary(u)?));
         context.expect_set_transient_storage().return_const(());
 
-        let revision = u
-            .choose(&[
-                Revision::EVMC_FRONTIER,
-                Revision::EVMC_HOMESTEAD,
-                Revision::EVMC_TANGERINE_WHISTLE,
-                Revision::EVMC_SPURIOUS_DRAGON,
-                Revision::EVMC_BYZANTIUM,
-                Revision::EVMC_CONSTANTINOPLE,
-                Revision::EVMC_PETERSBURG,
-                Revision::EVMC_ISTANBUL,
-                Revision::EVMC_BERLIN,
-                Revision::EVMC_LONDON,
-                Revision::EVMC_PARIS,
-                Revision::EVMC_SHANGHAI,
-                Revision::EVMC_CANCUN,
-                Revision::EVMC_PRAGUE,
-                Revision::EVMC_OSAKA,
-            ])?
-            .clone();
+        let revision = *u.choose(&[
+            Revision::EVMC_FRONTIER,
+            Revision::EVMC_HOMESTEAD,
+            Revision::EVMC_TANGERINE_WHISTLE,
+            Revision::EVMC_SPURIOUS_DRAGON,
+            Revision::EVMC_BYZANTIUM,
+            Revision::EVMC_CONSTANTINOPLE,
+            Revision::EVMC_PETERSBURG,
+            Revision::EVMC_ISTANBUL,
+            Revision::EVMC_BERLIN,
+            Revision::EVMC_LONDON,
+            Revision::EVMC_PARIS,
+            Revision::EVMC_SHANGHAI,
+            Revision::EVMC_CANCUN,
+            Revision::EVMC_PRAGUE,
+            Revision::EVMC_OSAKA,
+        ])?;
         let args = Self {
             instance: Instance::default(),
             host: mocked_host_interface(),
@@ -226,6 +211,6 @@ fuzz_target!(|args: InterpreterArgs| {
         &mut args.context,
         args.revision,
         &args.message,
-        &args.code,
+        args.code,
     );
 });