--- a/rust/src/utils/helpers.rs
+++ b/rust/src/utils/helpers.rs
@@ -1,9 +1,8 @@
 use std::cmp::min;
 
-use evmc_vm::{MessageFlags, Revision};
+use evmc_vm::{ExecutionMessage, MessageFlags, Revision};
 
 use crate::{
-    interpreter::Interpreter,
     types::{u256, FailStatus},
     utils::Gas,
 };
@@ -61,8 +60,8 @@ pub fn check_min_revision(min_revision: Revision, revision: Revision) -> Result<
 }
 
 #[inline(always)]
-pub fn check_not_read_only(state: &Interpreter) -> Result<(), FailStatus> {
-    if state.message.flags() == MessageFlags::EVMC_STATIC as u32 {
+pub fn check_not_read_only(message: &ExecutionMessage) -> Result<(), FailStatus> {
+    if message.flags() == MessageFlags::EVMC_STATIC as u32 {
         return Err(FailStatus::StaticModeViolation);
     }
     Ok(())