This PR has the overarching goal of excluding the step check also when feature tail-call is enabled. In order to do so, instead of having various methods generic, now whole types are generic (Interpreter, OpFn, CodeAnalysis, ...). This makes it possible to call the correct version of Interpreter::next from an opcode impl, which is needed for tail-calls.

Other changes:
- merge const generics STEP_CHECK and JUMPDEST into a single const generic STEPPABLE
- implement Interpreter::new only for Interpreter<false> and Interpreter::new_steppable for Interpreter<true>. This avoids having to provide the generic argument explicitly and makes it impossible to misuse the api.

```
                   │ 2024-11-25T15:30#e4391e3#20-main/evmrs#performance │ 2024-11-27T09:42#20243ba#10/evmrs#performance │
                   │                       sec/op                       │       sec/op        vs base                   │
StaticOverhead/1/                                           2.544µ ± 0%          2.629µ ± 1%   +3.36% (p=0.000 n=20+10)
Inc/1/                                                      5.166µ ± 0%          5.103µ ± 1%   -1.23% (p=0.000 n=20+10)
Inc/10/                                                     5.165µ ± 0%          5.119µ ± 0%   -0.91% (p=0.001 n=20+10)
Fib/1/                                                      4.715µ ± 0%          4.500µ ± 1%   -4.55% (p=0.000 n=20+10)
Fib/5/                                                      15.04µ ± 1%          14.04µ ± 1%   -6.65% (p=0.000 n=20+10)
Fib/10/                                                     141.5µ ± 0%          128.7µ ± 0%   -9.02% (p=0.000 n=20+10)
Fib/15/                                                     1.343m ± 0%          1.230m ± 0%   -8.45% (p=0.000 n=20+10)
Fib/20/                                                     14.56m ± 0%          13.23m ± 0%   -9.12% (p=0.000 n=20+10)
Sha3/1/                                                     2.916µ ± 0%          2.848µ ± 1%   -2.33% (p=0.000 n=20+10)
Sha3/10/                                                    4.525µ ± 0%          4.266µ ± 0%   -5.73% (p=0.000 n=20+10)
Sha3/100/                                                   18.81µ ± 0%          18.18µ ± 1%   -3.36% (p=0.000 n=20+10)
Sha3/1000/                                                  191.9µ ± 0%          188.0µ ± 0%   -2.04% (p=0.000 n=20+10)
Arith/1/                                                    5.929µ ± 0%          5.750µ ± 1%   -3.01% (p=0.000 n=20+10)
Arith/10/                                                   12.32µ ± 0%          12.07µ ± 1%   -1.97% (p=0.000 n=20+10)
Arith/100/                                                  81.79µ ± 0%          73.53µ ± 1%  -10.10% (p=0.000 n=20+10)
Arith/280/                                                  232.9µ ± 0%          211.9µ ± 0%   -9.03% (p=0.000 n=20+10)
Memory/1/                                                   8.474µ ± 1%          7.980µ ± 2%   -5.83% (p=0.000 n=20+10)
Memory/10/                                                  14.91µ ± 1%          13.47µ ± 1%   -9.67% (p=0.000 n=20+10)
Memory/100/                                                 85.87µ ± 0%          73.02µ ± 0%  -14.96% (p=0.000 n=20+10)
Memory/1000/                                                717.4µ ± 0%          611.6µ ± 0%  -14.75% (p=0.000 n=20+10)
Memory/10000/                                               6.882m ± 0%          5.818m ± 0%  -15.46% (p=0.000 n=20+10)
Analysis/jumpdest/                                          2.533µ ± 0%          2.547µ ± 1%   +0.55% (p=0.011 n=20+10)
Analysis/stop/                                              2.531µ ± 1%          2.547µ ± 0%   +0.63% (p=0.034 n=20+10)
Analysis/push1/                                             2.526µ ± 0%          2.546µ ± 1%   +0.81% (p=0.005 n=20+10)
Analysis/push32/                                            2.529µ ± 1%          2.549µ ± 1%   +0.81% (p=0.000 n=20+10)
geomean                                                     28.28µ               26.74µ        -5.43%
```